macos/requirements.txt
# macOS Daemon Dependencies
pyyaml==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
psutil==5.9.6
watchdog==3.0.0
colorlog==6.8.0